Start of work for #127. 1. Added several new models to represent the job table + package import tables. 2. Added new repositories to access the job tables. 3. Added a new api call to get a list of background jobs for a feed 4. Added background jobs to the feed history view 5. Fix the packages page breaking if no tags are set on the nuget package 6. Reworked the import packages job to be simpler and allow itself to pick up from where it was if it restarts 7. Added a new import packages page which lets you view the progress of a current import job or an older job which has since finished

+++ NuFridge.Shared/Server/Web/Actions/NuGetApiV2/UploadPackageAction.cs
@@ -52,10 +52,14 @@ namespace NuFridge.Shared.Server.Web.Actions.NuGetApiV2
                 return errorResponse;
             }
 
+            _log.Warn("before claim");
+
             if (module.Context.CurrentUser != null && module.Context.CurrentUser.IsAuthenticated())
             {
                 module.RequiresAnyClaim(new List<string> { Claims.SystemAdministrator, Claims.CanUploadPackages });
             }
+
+            _log.Warn("after claim");
 
             if (RequiresApiKeyCheck(feed))
             {
@@ -77,15 +81,19 @@ namespace NuFridge.Shared.Server.Web.Actions.NuGetApiV2
                 }
                 else
                 {
-                    if (File.Exists(filePath))
+                    if (!IsValidNuGetApiKey(module, feed))
                     {
-                        File.Delete(filePath);
-                    }
+                        _log.Warn("Invalid API used to push package for feed " + feed.Name);
 
-                    _log.Warn("Invalid API used to push package for feed " + feed.Name);
-                    var errorResponse = module.Response.AsText("Invalid API key.");
-                    errorResponse.StatusCode = HttpStatusCode.Forbidden;
-                    return errorResponse;
+                        if (File.Exists(filePath))
+                        {
+                            File.Delete(filePath);
+                        }
+
+                        var errorResponse = module.Response.AsText("Invalid API key.");
+                        errorResponse.StatusCode = HttpStatusCode.Forbidden;
+                        return errorResponse;
+                    }
                 }
             }
